+++ core/execution/engine.go
@@ -221,6 +221,11 @@ func (e *Engine) ReloadConf(cfg Config) {
 	}
 }
 
+func (e *Engine) SpotsMarketsEnabled() bool {
+	// TODO replace with real implementation
+	return false
+}
+
 func (e *Engine) Hash() []byte {
 	e.log.Debug("hashing markets")
 
@@ -330,8 +335,8 @@ func (e *Engine) SucceedMarket(ctx context.Context, successor, parent string, in
 	}
 	// we have an actual parent market and some state, so this is a "true" successor market.
 	if pMarket != nil {
-		asset, _ := pMarket.GetAsset()
-		lm := e.collateral.SuccessorInsuranceFraction(ctx, successor, parent, asset, insuranceFraction)
+		assets, _ := pMarket.GetAssets()
+		lm := e.collateral.SuccessorInsuranceFraction(ctx, successor, parent, assets[0], insuranceFraction)
 		e.broker.Send(events.NewLedgerMovements(ctx, []*types.LedgerMovement{lm}))
 		// now we can set the ELS stuff
 		sMkt.InheritParent(ctx, cpState)
@@ -364,11 +369,7 @@ func (e *Engine) IsEligibleForProposerBonus(marketID string, value *num.Uint) bo
 	if _, ok := e.markets[marketID]; !ok {
 		return false
 	}
-	asset, err := e.markets[marketID].mkt.GetAsset()
-	if err != nil {
-		return false
-	}
-	quantum, err := e.collateral.GetAssetQuantum(asset)
+	quantum, err := e.collateral.GetAssetQuantum(e.markets[marketID].settlementAsset)
 	if err != nil {
 		return false
 	}
@@ -417,10 +418,11 @@ func (e *Engine) submitOrRestoreMarket(ctx context.Context, marketConfig *types.
 	}
 
 	if isNewMarket {
-		asset, err := marketConfig.GetAsset()
+		assets, err := marketConfig.GetAssets()
 		if err != nil {
 			e.log.Panic("failed to get asset from market config", logging.String("market", marketConfig.ID), logging.String("error", err.Error()))
 		}
+		asset := assets[0]
 		e.marketActivityTracker.MarketProposed(asset, marketConfig.ID, proposer)
 	}
 
@@ -468,10 +470,12 @@ func (e *Engine) submitMarket(ctx context.Context, marketConfig *types.Market, p
 	now := e.timeService.GetTimeNow()
 
 	// ensure the asset for this new market exists
-	asset, err := marketConfig.GetAsset()
+	assets, err := marketConfig.GetAssets()
 	if err != nil {
 		return err
 	}
+	asset := assets[0]
+
 	if !e.collateral.AssetExists(asset) {
 		e.log.Error("unable to create a market with an invalid asset",
 			logging.MarketID(marketConfig.ID),