+++ core/governance/engine.go
@@ -45,6 +45,7 @@ var (
 	ErrUnsupportedAssetSourceType                = errors.New("unsupported asset source type")
 	ErrExpectedERC20Asset                        = errors.New("expected an ERC20 asset but was not")
 	ErrErc20AddressAlreadyInUse                  = errors.New("erc20 address already in use")
+	ErrSpotsNotEnabled                           = errors.New("spot trading not enabled")
 	ErrParentMarketDoesNotExist                  = errors.New("market to succeed does not exist")
 	ErrParentMarketAlreadySucceeded              = errors.New("the market was already succeeded by a prior proposal")
 )
@@ -67,6 +68,7 @@ type Markets interface {
 	RestoreMarket(ctx context.Context, marketConfig *types.Market) error
 	StartOpeningAuction(ctx context.Context, marketID string) error
 	UpdateMarket(ctx context.Context, marketConfig *types.Market) error
+	SpotsMarketsEnabled() bool
 }
 
 // StakingAccounts ...
@@ -227,12 +229,20 @@ func (e *Engine) preEnactProposal(ctx context.Context, p *proposal) (te *ToEnact
 	switch p.Terms.Change.GetTermType() {
 	case types.ProposalTermsTypeNewMarket:
 		te.m = &ToEnactNewMarket{}
+	case types.ProposalTermsTypeNewSpotMarket:
+		te.s = &ToEnactNewSpotMarket{}
 	case types.ProposalTermsTypeUpdateMarket:
 		mkt, perr, err := e.updatedMarketFromProposal(p)
 		if err != nil {
 			return nil, perr, err
 		}
 		te.updatedMarket = mkt
+	case types.ProposalTermsTypeUpdateSpotMarket:
+		mkt, perr, err := e.updatedSpotMarketFromProposal(p)
+		if err != nil {
+			return nil, perr, err
+		}
+		te.updatedSpotMarket = mkt
 	case types.ProposalTermsTypeUpdateNetworkParameter:
 		unp := p.Terms.GetUpdateNetworkParameter()
 		if unp != nil {
@@ -592,6 +602,27 @@ func (e *Engine) intoToSubmit(ctx context.Context, p *types.Proposal, enct *enac
 		tsb.m = &ToSubmitNewMarket{
 			m: mkt,
 		}
+	case types.ProposalTermsTypeNewSpotMarket:
+		closeTime := e.timeService.GetTimeNow().Truncate(time.Second)
+		enactTime := time.Unix(p.Terms.EnactmentTimestamp, 0)
+		newMarket := p.Terms.GetNewSpotMarket()
+		auctionDuration := enactTime.Sub(closeTime)
+		if !e.markets.SpotsMarketsEnabled() {
+			e.rejectProposal(ctx, p, types.ProposalErrorSpotNotEnabled, ErrSpotsNotEnabled)
+			return nil, fmt.Errorf("%w, %v", ErrSpotsNotEnabled, types.ProposalErrorSpotNotEnabled)
+		}
+		if perr, err := validateNewSpotMarketChange(newMarket, e.assets, true, e.netp, auctionDuration, enct); err != nil {
+			e.rejectProposal(ctx, p, perr, err)
+			return nil, fmt.Errorf("%w, %v", err, perr)
+		}
+		mkt, perr, err := buildSpotMarketFromProposal(p.ID, newMarket, e.netp, auctionDuration)
+		if err != nil {
+			e.rejectProposal(ctx, p, perr, err)
+			return nil, fmt.Errorf("%w, %v", err, perr)
+		}
+		tsb.s = &ToSubmitNewSpotMarket{
+			m: mkt,
+		}
 	}
 
 	return tsb, nil
@@ -642,6 +673,10 @@ func (e *Engine) getProposalParams(terms *types.ProposalTerms) (*ProposalParamet
 		return e.getUpdateNetworkParameterProposalParameters(), nil
 	case types.ProposalTermsTypeNewFreeform:
 		return e.getNewFreeformProposalParameters(), nil
+	case types.ProposalTermsTypeNewSpotMarket:
+		return e.getNewSpotMarketProposalParameters(), nil
+	case types.ProposalTermsTypeUpdateSpotMarket:
+		return e.getUpdateSpotMarketProposalParameters(), nil
 	default:
 		return nil, ErrUnsupportedProposalType
 	}
@@ -751,6 +786,13 @@ func (e *Engine) validateOpenProposal(proposal *types.Proposal) (types.ProposalE
 
 	if proposal.IsMarketUpdate() {
 		proposalError, err := e.validateMarketUpdate(proposal, params)
+		if err != nil {
+			return proposalError, err
+		}
+	}
+
+	if proposal.IsSpotMarketUpdate() {
+		proposalError, err := e.validateSpotMarketUpdate(proposal, params)
 		if err != nil {
 			return proposalError, err
 		}
@@ -874,6 +916,36 @@ func (e *Engine) validateMarketUpdate(proposal *types.Proposal, params *Proposal
 	return types.ProposalErrorUnspecified, nil
 }
 
+func (e *Engine) validateSpotMarketUpdate(proposal *types.Proposal, params *ProposalParameters) (types.ProposalError, error) {
+	updateMarket := proposal.SpotMarketUpdate()
+	if !e.markets.MarketExists(updateMarket.MarketID) {
+		e.log.Debug("market does not exist",
+			logging.MarketID(updateMarket.MarketID),
+			logging.PartyID(proposal.Party),
+			logging.ProposalID(proposal.ID))
+		return types.ProposalErrorInvalidMarket, ErrMarketDoesNotExist
+	}
+	for _, p := range e.activeProposals {
+		if p.ID == updateMarket.MarketID {
+			return types.ProposalErrorInvalidMarket, ErrMarketNotEnactedYet
+		}
+	}
+
+	partyELS, _ := e.markets.GetEquityLikeShareForMarketAndParty(updateMarket.MarketID, proposal.Party)
+	if partyELS.LessThan(params.MinEquityLikeShare) {
+		e.log.Debug("proposer have insufficient equity-like share",
+			logging.String("expect-balance", params.MinEquityLikeShare.String()),
+			logging.String("proposer-balance", partyELS.String()),
+			logging.PartyID(proposal.Party),
+			logging.MarketID(updateMarket.MarketID),
+			logging.ProposalID(proposal.ID))
+		return types.ProposalErrorInsufficientEquityLikeShare,
+			fmt.Errorf("proposer have insufficient equity-like share, expected >= %v got %v", params.MinEquityLikeShare, partyELS)
+	}
+
+	return types.ProposalErrorUnspecified, nil
+}
+
 // validates proposed change.
 func (e *Engine) validateChange(terms *types.ProposalTerms) (types.ProposalError, error) {
 	enactTime := time.Unix(terms.EnactmentTimestamp, 0)
@@ -901,6 +973,15 @@ func (e *Engine) validateChange(terms *types.ProposalTerms) (types.ProposalError
 		return terms.GetUpdateAsset().Validate()
 	case types.ProposalTermsTypeUpdateNetworkParameter:
 		return validateNetworkParameterUpdate(e.netp, terms.GetUpdateNetworkParameter().Changes)
+	case types.ProposalTermsTypeNewSpotMarket:
+		if !e.markets.SpotsMarketsEnabled() {
+			return types.ProposalErrorSpotNotEnabled, ErrSpotsNotEnabled
+		}
+		closeTime := time.Unix(terms.ClosingTimestamp, 0)
+		return validateNewSpotMarketChange(terms.GetNewSpotMarket(), e.assets, true, e.netp, enactTime.Sub(closeTime), enct)
+	case types.ProposalTermsTypeUpdateSpotMarket:
+		enct.shouldNotVerify = true
+		return validateUpdateSpotMarketChange(terms.GetUpdateSpotMarket())
 	default:
 		return types.ProposalErrorUnspecified, nil
 	}
@@ -1002,6 +1083,31 @@ func (e *Engine) updateValidatorKey(ctx context.Context, m map[string]*types.Vot
 	}
 }
 
+func (e *Engine) updatedSpotMarketFromProposal(p *proposal) (*types.Market, types.ProposalError, error) {
+	terms := p.Terms.GetUpdateSpotMarket()
+	existingMarket, exists := e.markets.GetMarket(terms.MarketID, false)
+	if !exists {
+		return nil, types.ProposalErrorInvalidMarket, fmt.Errorf("market \"%s\" doesn't exist anymore", terms.MarketID)
+	}
+
+	newMarket := &types.NewSpotMarket{
+		Changes: &types.NewSpotMarketConfiguration{
+			DecimalPlaces:             existingMarket.DecimalPlaces,
+			PositionDecimalPlaces:     existingMarket.PositionDecimalPlaces,
+			Metadata:                  terms.Changes.Metadata,
+			PriceMonitoringParameters: terms.Changes.PriceMonitoringParameters,
+			TargetStakeParameters:     terms.Changes.TargetStakeParameters,
+		},
+	}
+
+	if perr, err := validateUpdateSpotMarketChange(terms); err != nil {
+		return nil, perr, err
+	}
+
+	previousAuctionDuration := time.Duration(existingMarket.OpeningAuction.Duration) * time.Second
+	return buildSpotMarketFromProposal(existingMarket.ID, newMarket, e.netp, previousAuctionDuration)
+}
+
 func (e *Engine) updatedMarketFromProposal(p *proposal) (*types.Market, types.ProposalError, error) {
 	terms := p.Terms.GetUpdateMarket()
 	existingMarket, exists := e.markets.GetMarket(terms.MarketID, false)
@@ -1045,10 +1151,10 @@ func (e *Engine) updatedMarketFromProposal(p *proposal) (*types.Market, types.Pr
 	case nil:
 		return nil, types.ProposalErrorNoProduct, ErrMissingProduct
 	case *types.UpdateInstrumentConfigurationFuture:
-		asset, _ := existingMarket.GetAsset()
+		assets, _ := existingMarket.GetAssets()
 		newMarket.Changes.Instrument.Product = &types.InstrumentConfigurationFuture{
 			Future: &types.FutureProduct{
-				SettlementAsset:                     asset,
+				SettlementAsset:                     assets[0],
 				QuoteName:                           product.Future.QuoteName,
 				DataSourceSpecForSettlementData:     product.Future.DataSourceSpecForSettlementData,
 				DataSourceSpecForTradingTermination: product.Future.DataSourceSpecForTradingTermination,