+++ core/governance/mocks/mocks.go
@@ -115,6 +115,20 @@ func (mr *MockMarketsMockRecorder) RestoreMarket(arg0, arg1 interface{}) *gomock
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "RestoreMarket", reflect.TypeOf((*MockMarkets)(nil).RestoreMarket), arg0, arg1)
 }
 
+// SpotsMarketsEnabled mocks base method.
+func (m *MockMarkets) SpotsMarketsEnabled() bool {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "SpotsMarketsEnabled")
+	ret0, _ := ret[0].(bool)
+	return ret0
+}
+
+// SpotsMarketsEnabled indicates an expected call of SpotsMarketsEnabled.
+func (mr *MockMarketsMockRecorder) SpotsMarketsEnabled() *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SpotsMarketsEnabled", reflect.TypeOf((*MockMarkets)(nil).SpotsMarketsEnabled))
+}
+
 // StartOpeningAuction mocks base method.
 func (m *MockMarkets) StartOpeningAuction(arg0 context.Context, arg1 string) error {
 	m.ctrl.T.Helper()