+++ core/integration/main_test.go
@@ -175,8 +175,8 @@ func InitializeScenario(s *godog.ScenarioContext) {
 	s.Step(`^the initial insurance pool balance is "([^"]*)" for all the markets$`, func(amountstr string) error {
 		amount, _ := num.UintFromString(amountstr, 10)
 		for _, mkt := range execsetup.markets {
-			asset, _ := mkt.GetAsset()
-			marketInsuranceAccount, err := execsetup.collateralEngine.GetMarketInsurancePoolAccount(mkt.ID, asset)
+			assets, _ := mkt.GetAssets()
+			marketInsuranceAccount, err := execsetup.collateralEngine.GetMarketInsurancePoolAccount(mkt.ID, assets[0])
 			if err != nil {
 				return err
 			}