+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - [8283](https://github.com/vegaprotocol/vega/issues/8283) - Add disclaimer to the wallet `CLI`
 - [8296](https://github.com/vegaprotocol/vega/issues/8296) - Improve error handling for invalid proposal validation timestamp
 - [8318](https://github.com/vegaprotocol/vega/issues/8318) - Proto definitions for spots
+- [8117](https://github.com/vegaprotocol/vega/issues/8117) - Added spots governance implementation
 - [8259](https://github.com/vegaprotocol/vega/issues/8259) - Proto definitions for successor markets.
 - [8201](https://github.com/vegaprotocol/vega/issues/8201) - Add support for successor markets.